# Conflicts:
#	src/main/java/converter/BusConverter.java
#	src/main/java/converter/BusStopConverter.java
#	src/main/java/converter/BusStopLineConverter.java
#	src/main/java/converter/BusStopPathConverter.java
#	src/main/java/converter/LineConverter.java
#	src/main/java/converter/PathConverter.java
#	src/main/java/converter/PersonConverter.java
#	src/main/java/converter/WarehouseConverter.java
#	src/main/java/dao/BusDAO.java
#	src/main/java/dao/BusStopDAO.java
#	src/main/java/dao/LineDAO.java
#	src/main/java/dao/PersonDAO.java
#	src/main/java/modele/BusStop.java
#	src/main/java/modele/BusStopLine.java
#	src/main/java/modele/BusStopPath.java
#	src/main/java/modele/Line.java
#	src/main/java/modele/Path.java
#	src/main/java/modele/Person.java
#	src/main/java/services/Services.java
#	src/main/java/servlets/ActionServlet.java

+++ src/main/java/services/Algorithm.java
@@ -14,8 +14,8 @@ import modele.*;
  * @author
  */
 public class Algorithm {
-    private static float[][] durations;
-    private static float pourcentage;
+    private static double[][] durations;
+    private static double pourcentage;
     private static Date currentDate;
     
     public static double getCost(ArrayList<ArrayList<Person>> journeys, ArrayList<Line> lines){
@@ -69,7 +69,7 @@ public class Algorithm {
         return cost/2;      //because we calculate twice for each person
     }
     
-    public static ArrayList<Line> calculateLines (float[][]journeyDurations, Bus[]buses, ArrayList <Person> requests, Date theCurrentDate){
+    public static ArrayList<Line> calculateLines (double[][]journeyDurations, Bus[]buses, ArrayList <Person> requests, Date theCurrentDate){
         durations = journeyDurations;
         currentDate = theCurrentDate;
         pourcentage = 3;
